refactoring
separate the code that’s specific to our schema into models.js and
createRootValue in graphql/index.js; everything else should now be
reusable for other bookshelf schemas
